# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "AquaTracker app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of AquaTracker app\n"
  },
  "servers": [
    {
      "url": "https://aquatracker-node.onrender.com"
    }
  ],
...
  "paths": {
    "/users/register": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Register a new user",
        "operationId": "registerUser",
        "description": "Register a new user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "description": "User email",
                    "type": "string",
                    "example": "bob@me.com"
                  },
                  "password": {
                    "description": "User password",
                    "type": "string",
                    "example": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully registered a user!",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully registered a user!"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "description": "New user id (auto-generated)",
                          "type": "string",
                          "example": "666ed33db5653b741ac1b547"
                        },
                        "name": {
                          "description": "New user name",
                          "type": "string",
                          "default": "User",
                          "example": "Robert Pattison"
                        },
                        "email": {
                          "description": "New user email",
                          "type": "string",
                          "example": "bob@me.com"
                        },
                        "gender": {
                          "type": "string",
                          "enum": [
                            "man",
                            "woman"
                          ],
                          "default": "woman"
                        },
                        "dailyNorma": {
                          "type": "number",
                          "default": 1500
                        },
                        "weight": {
                          "type": "number",
                          "default": 0
                        },
                        "activeHours": {
                          "type": "number",
                          "default": 0
                        },
                        "avatar": {
                          "type": "string",
                          "format": "binary",
                          "default": "Here should be a placeholder image"
                        },
                        "createdAt": {
                          "description": "Timestamp (auto-generated)",
                          "type": "string",
                          "example": "2024-06-16T11:57:49.764Z"
                        },
                        "updatedAt": {
                          "description": "Timestamp (auto-generated)",
                          "type": "string",
                          "example": "2024-06-16T11:57:49.764Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "409": {
            "description": "Email is already in use",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 409
                    },
                    "message": {
                      "type": "string",
                      "example": "ConflictError"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Email is already in use"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Login user",
        "operationId": "loginUser",
        "description": "Login user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "description": "User email",
                    "type": "string",
                    "example": "bob@me.com"
                  },
                  "password": {
                    "description": "User password",
                    "type": "string",
                    "example": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully logged in an user!",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully logged in an user!"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "accessToken": {
                          "type": "string",
                          "example": "EyShUQ68S0AK8fL0+6g/HuLYPTZ+so1KUHyRZCwk"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    },
    "/users/refresh": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Resfresh user session",
        "operationId": "refreshUserSession",
        "description": "Resfresh user session",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "refreshToken",
            "in": "cookie",
            "required": true,
            "schema": {
              "description": "Refresh token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully refreshed a session!",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully refreshed a session!"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "accessToken": {
                          "type": "string",
                          "example": "+ktMRdCpqUs482YQY57TK6Z5ME8u75fZYQYFfNks"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Session token expired",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Session token expired"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Session token expired"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/update": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Updates user profile details and preferences.",
        "operationId": "updateUserController",
        "description": "Update user account.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "description": "User email",
                    "type": "string",
                    "example": "user@email.com"
                  },
                  "name": {
                    "description": "User name",
                    "type": "string",
                    "example": "2025-03-01",
                    "default": "User"
                  },
                  "gender": {
                    "type": "string",
                    "enum": [
                      "man",
                      "woman"
                    ],
                    "default": "woman"
                  },
                  "dailyNorma": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 10000,
                    "default": 1500
                  },
                  "weight": {
                    "type": "number",
                    "default": 0
                  },
                  "activeHours": {
                    "type": "number",
                    "default": 0
                  },
                  "avatar": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request successfully completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully updated user"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/user"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/users/logout": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Logout user",
        "operationId": "logoutUser",
        "description": "Logout user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "No Content - User was successfully logged out"
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    },
    "/users/current": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Gets information about current user.",
        "operationId": "getCurrentUserController",
        "description": "Get info about current user.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Request successfully completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully retrieved user information!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/user"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "responses": {
      "400": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "Something went wrong"
                },
                "error": {
                  "type": "string",
                  "example": "Bad request"
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access token expired"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "user": {
        "title": "User schema",
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "_id": {
            "description": "User id assigned by MongoDB.",
            "type": "string",
            "example": "65e4decdd286b30065d54af9"
          },
          "email": {
            "description": "User email",
            "type": "string",
            "example": "user@email.com"
          },
          "name": {
            "description": "User name",
            "type": "string",
            "example": "Valeriy Zaluzhnyi",
            "default": "User"
          },
          "password": {
            "description": "User password"
          },
          "gender": {
            "type": "string",
            "enum": [
              "man",
              "woman"
            ],
            "default": "woman"
          },
          "dailyNorma": {
            "type": "number",
            "minimum": 0,
            "maximum": 10000,
            "default": 1500
          },
          "weight": {
            "type": "number",
            "default": 0
          },
          "activeHours": {
            "type": "number",
            "default": 0
          },
          "avatar": {
            "type": "string",
            "format": "binary"
          },
          "createdAt": {
            "description": "User account creation date assigned by MongoDB.",
            "type": "string",
            "example": "2024-07-01T18:49:47.392Z"
          },
          "updatedAt": {
            "description": "Last time user account was updated assigned by MongoDB.",
            "type": "string",
            "example": "2024-07-01T18:49:47.392Z"
          }
        }
      },
...
    }
  }
}